chore: improve Stack exports
